pyproject.toml
[tool.poetry]
name = "aider-runtime"
version = "0.1.0"
description = "Aider runtime"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.*"
aider-chat = {extras = ["browser"], version = "*"}
google-cloud-aiplatform = "*"
